Added in Normal-Rayleigh model
Added option to estimate a Normal-Rayleigh model using the option "NR" in the sfm function. Seems to be working fine.

N.B. I have used a slightly different parameterisation to that usually used for the Rayleigh distribution (and used in Hajargasht's paper). The parameterisation I have used is that which arises if we start with a Nakagami distribution and set the shape parameter to 2. This is in order to make comparison easier -- I plan to add in the Nakagami model soon.

(this version should fix previous conflicts)